+++ lib/ig_markets/cli/commands/activities_command.rb
@@ -6,6 +6,7 @@ module IGMarkets
 
       option :days, type: :numeric, required: true, desc: 'The number of days to print account activities for'
       option :from, desc: 'The start date to print account activities from, format: yyyy-mm-dd'
+      option :epic, desc: 'Regex for filtering activities based on their EPIC'
       option :sort_by, enum: %w(channel date epic type), default: 'date', desc: 'The attribute to sort activities by'
 
       def activities
@@ -21,9 +22,19 @@ module IGMarkets
       private
 
       def gather_activities(dealing_platform)
-        gather_account_history(:activities, dealing_platform).sort_by do |activity|
+        @epic_regex = Regexp.new options.fetch('epic', ''), Regexp::IGNORECASE
+
+        result = gather_account_history(:activities, dealing_platform).select do |activity|
+          activity_filter activity
+        end
+
+        result.sort_by do |activity|
           [activity.send(activity_sort_attribute), activity.date]
         end
+      end
+
+      def activity_filter(activity)
+        @epic_regex.match activity.epic
       end
 
       def activity_sort_attribute